feat: change path to v1
