+++ internal/server/http/handler/v1/error_handler/error_handler.go
@@ -1,4 +1,4 @@
-package handler
+package error_handler
 
 import (
 	"errors"
