+++ internal/server/http/handler/v1/info/handler.go
@@ -4,14 +4,14 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/green-ecolution/green-ecolution-backend/internal/server/http/entities/info"
 	"github.com/green-ecolution/green-ecolution-backend/internal/server/http/entities/info/generated"
-	"github.com/green-ecolution/green-ecolution-backend/internal/server/http/handler"
+	"github.com/green-ecolution/green-ecolution-backend/internal/server/http/handler/v1/error_handler"
 	"github.com/green-ecolution/green-ecolution-backend/internal/service"
 )
 
 // @Summary		Get info about the app
 // @Description	Get info about the app and the server
 // @Id				get-app-info
-// @Tags			Info
+// @Tags			v1,Info
 // @Produce		json
 // @Success		200	{object}	info.AppInfoResponse
 // @Failure		400	{object}	HTTPError
@@ -19,14 +19,14 @@ import (
 // @Failure		403	{object}	HTTPError
 // @Failure		404	{object}	HTTPError
 // @Failure		500	{object}	HTTPError
-// @Router			/info [get]
+// @Router			/v1/info [get]
 func GetAppInfo(svc service.InfoService) fiber.Handler {
 	var mapper info.InfoHTTPMapper = &generated.InfoHTTPMapperImpl{}
 
 	return func(c *fiber.Ctx) error {
 		domainInfo, err := svc.GetAppInfoResponse(c.Context())
 		if err != nil {
-			return handler.HandleError(err)
+			return error_handler.HandleError(err)
 		}
 
 		response := mapper.ToResponse(domainInfo)
